Move tests into separate package
