+++ crates/bpe-openai/Cargo.toml
@@ -20,7 +20,7 @@ rmp-serde = "1"
 serde = { version = "1" }
 
 [dev-dependencies]
-tiktoken-rs = { version = "0.6" }
+tiktoken-rs = "0.6"
 
 [build-dependencies]
 base64 = "0.22.1"
